% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.mars.R
\name{mice.impute.mars}
\alias{mice.impute.mars}
\title{MARS algorithm to be used within \code{mice}}
\usage{
mice.impute.mars(y, ry, x, wy = NULL, degree = 10, sampling = 0, ...)
}
\arguments{
\item{y}{Vector to be imputed.}

\item{ry}{Logical vector of length \code{length(y)} indicating the the subset \code{y[ry]} of elements in y to which the imputation model is fitted. The \code{ry} generally distinguishes the observed (\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for \code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value indicates locations in \code{y} for which imputations are created.}

\item{degree}{Maximum degree of interaction. Default is 10.}

\item{sampling}{Amount of subsampling that is applied. Default is 0 which corresponds to bootstrap resampling.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length \code{sum(wy)}
}
\description{
Imputes univariate missing data using the MARS algorithm with subsampling.
}
\examples{
# We normally call mice.impute.mars() from within mice()
}
